Doo

See merge request wkzkfmxk12/invest-backend!156

+++ microservices/investment/src/main/java/com/illunex/invest/investment/service/EvaluateService.java
@@ -85,21 +85,20 @@ public class EvaluateService {
         } else {
 
             for (EvaluateReviewItemCategory c : newEvaluate.getTemplate().getReviewItemCategory()) {
-
                 for (EvaluateReviewItem i : c.getReviewItem()) {
                     evaluateReviewItemRepository.deleteAllByReviewItemCategoryIdx(c.getIdx());
                     i.setReviewItemCategory(c);
                 }
-
                 evaluateReviewItemCategoryRepository.deleteAllByReviewItemTemplateIdx(newEvaluate.getTemplate().getIdx());
                 c.setReviewItemTemplate(newEvaluate.getTemplate());
-
             }
 
             for (EvaluateJudge j : newEvaluate.getJudgeList()) {
-                for (EvaluateJudgeScore s : j.getScoreList()) {
-                    evaluateJudgeScoreRepository.deleteAllByJudgeIdx(j.getIdx());
-                    s.setJudge(j);
+                if(j.getDivision().equals("attend")) {
+                    for (EvaluateJudgeScore s : j.getScoreList()) {
+                        evaluateJudgeScoreRepository.deleteAllByJudgeIdx(j.getIdx());
+                        s.setJudge(j);
+                    }
                 }
                 evaluateJudgeRepository.deleteAllByEvaluateIdx(newEvaluate.getIdx());
                 j.setEvaluate(evaluate);
@@ -116,16 +115,20 @@ public class EvaluateService {
             Evaluate editingEvaluate = evaluateRepository.save(evaluate);
 
             for (EvaluateJudge j : newEvaluate.getJudgeList()) {
-                for (EvaluateReviewItemCategory c : editingEvaluate.getTemplate().getReviewItemCategory()) {
-                    for (EvaluateReviewItem i : c.getReviewItem()) {
-                        j.getScoreList().add(EvaluateJudgeScore.builder()
-                            .categoryIdx(c.getIdx())
-                            .reviewItemIdx(i.getIdx())
-                            .score(0)
-                            .judge(j)
-                            .build());
+
+                if(j.getDivision().equals("attend")) {
+                    for (EvaluateReviewItemCategory c : editingEvaluate.getTemplate().getReviewItemCategory()) {
+                        for (EvaluateReviewItem i : c.getReviewItem()) {
+                            j.getScoreList().add(EvaluateJudgeScore.builder()
+                                .categoryIdx(c.getIdx())
+                                .reviewItemIdx(i.getIdx())
+                                .score(0)
+                                .judge(j)
+                                .build());
+                        }
                     }
                 }
+
             }
 
             evaluateRepository.save(editingEvaluate);
@@ -140,14 +143,32 @@ public class EvaluateService {
         if (evaluate == null) {
             return "Invalid Evaluate";
         } else {
+            int attendCount = 0;
+            int completeCount = 0;
+
             for (EvaluateJudge j : evaluate.getJudgeList()) {
-                if (j.getUserIdx().equals(evaluateCommentDTO.getUserIdx())) {
+                if (j.getDivision().equals("attend")) attendCount++;
+                if ( (j.getUserIdx().equals(evaluateCommentDTO.getUserIdx())) && (j.getDivision().equals("attend")) )  {
                     j.setComment(evaluateCommentDTO.getComment());
                     j.setStatus("complete");
                     j.setEvaluateDate(LocalDateTime.now());
+                    completeCount++;
                 }
+                if (j.getStatus().equals("complete")) completeCount++;
             }
-            evaluateRepository.save(evaluate);
+            Evaluate editingEvaluate = evaluateRepository.save(evaluate);
+
+            if (completeCount == attendCount) {
+                float judgeTotalScore = 0;
+                for (EvaluateJudge e: editingEvaluate.getJudgeList()) {
+                    judgeTotalScore += e.getFinalScore();
+                }
+                editingEvaluate.setStatus("complete");
+                editingEvaluate.setAverageScore((int) (judgeTotalScore/completeCount));
+                editingEvaluate.setCompleteDate(LocalDateTime.now());
+                evaluateRepository.save(editingEvaluate);
+            }
+
             return "Evaluate complete";
         }
     }
@@ -177,11 +198,10 @@ public class EvaluateService {
             return "Invalid Evaluate";
         } else {
             EvaluateJudge judge = mapper.evaluateJudgeDTOToEntity(evaluateReviewDTO.getJudge());
-            int completeCount = 0;
 
             for (EvaluateJudge e: evaluate.getJudgeList()) {
-                if (e.getUserIdx().equals(evaluateReviewDTO.getJudge().getUserIdx())) {
-                    e.setStatus("confirm");
+                if ( (e.getUserIdx().equals(evaluateReviewDTO.getJudge().getUserIdx())) && (e.getDivision().equals("attend")) )  {
+                    e.setStatus("evaluated");
 
                     for (EvaluateJudgeScore j : e.getScoreList()) {
                         for (EvaluateJudgeScore s : judge.getScoreList()) {
@@ -213,22 +233,8 @@ public class EvaluateService {
 
                     e.setFinalScore((int)judgeScore);
                 }
-
-                if (e.getStatus().equals("complete")) completeCount++;
             }
 
-
-            if (completeCount == evaluate.getJudgeList().size()) {
-                float judgeTotalScore = 0;
-
-                for (EvaluateJudge e: evaluate.getJudgeList()) {
-                    judgeTotalScore += e.getFinalScore();
-                }
-
-                evaluate.setStatus("complete");
-                evaluate.setAverageScore((int) (judgeTotalScore/3));
-                evaluate.setCompleteDate(LocalDateTime.now());
-            }
             evaluateRepository.save(evaluate);
             return "Evaluate Review "+ evaluate.getStatus();
         }
